Feature: support for i386 Linux system calls.
* Added i386 kernel calling convention support.
* C parser support for new [[reko::characteristics]] attribute
* Fix #1147

+++ src/Decompiler/Scanning/BlockWorkitem.cs
@@ -1274,8 +1274,8 @@ namespace Reko.Scanning
 
             if (!(fn.Arguments[0] is Constant vector))
                 return null;
-            var svc = program.Platform.FindService(vector.ToInt32(), state, program.SegmentMap);
-            //$TODO if SVC uis null (and not-speculating) report the error.
+            var svc = program.Platform.FindService((int)vector.ToUInt32(), state, program.SegmentMap);
+            //$TODO if SVC is null (and not-speculating) report the error.
             if (svc != null && svc.Signature == null)
             {
                 scanner.Error(ric!.Address, string.Format("System service '{0}' didn't specify a signature.", svc.Name));
